userspace/nat: split NPTv6 compiler into nat_nptv6.go
Final increment of the #4421 (codex-173 #5) pure code-motion split.
Moves the NPTv6 declarations verbatim out of nat.go into nat_nptv6.go:
buildNptv6Snapshots (the IPv6 prefix-translation snapshot builder that
walks the static rule-sets for IsNPTv6 rules) and hasNonNptv6StaticNAT
(the companion predicate reporting whether any non-NPTv6 static rule
exists).

After this move nat.go is the ~225-line shared residual: natCounterID,
the NAT address-name resolvers (resolveNATAddressNamePrefixes,
appendNATSource/DestinationAddressName), coalescePortRanges,
appPortsFromSpec, and the natNeverMatchPortRange sentinel — the helpers
used across two or more NAT families. Each family now lives in its own
sibling file (nat_source.go, nat_static.go, nat_destination.go,
nat64.go, nat_nptv6.go).

No declaration body was edited across the whole split and nothing was
renamed. The sorted top-level production decl-name set is byte-identical
before and after (empty diff), and the non-blank declaration-body line
multiset is identical (1233 lines).

Validation: go build ./... green, go vet clean, go test
./pkg/dataplane/userspace/... green.
