+++ pkg/dataplane/userspace/nat_nptv6.go
@@ -1,0 +1,48 @@
+package userspace
+
+import (
+	"github.com/psaab/xpf/pkg/config"
+)
+
+func buildNptv6Snapshots(cfg *config.Config) []Nptv6RuleSnapshot {
+	if cfg == nil || len(cfg.Security.NAT.Static) == 0 {
+		return nil
+	}
+	var out []Nptv6RuleSnapshot
+	for _, rs := range cfg.Security.NAT.Static {
+		if rs == nil {
+			continue
+		}
+		for _, rule := range rs.Rules {
+			if rule == nil || !rule.IsNPTv6 {
+				continue
+			}
+			out = append(out, Nptv6RuleSnapshot{
+				Name:           rule.Name,
+				FromZone:       rs.FromZone,
+				ExternalPrefix: rule.Match,
+				InternalPrefix: rule.Then,
+			})
+		}
+	}
+	return out
+}
+
+// hasNonNptv6StaticNAT returns true if the config has any static NAT rules
+// that are NOT NPTv6. NPTv6 rules are supported by the userspace dataplane.
+func hasNonNptv6StaticNAT(cfg *config.Config) bool {
+	if cfg == nil {
+		return false
+	}
+	for _, rs := range cfg.Security.NAT.Static {
+		if rs == nil {
+			continue
+		}
+		for _, rule := range rs.Rules {
+			if rule != nil && !rule.IsNPTv6 {
+				return true
+			}
+		}
+	}
+	return false
+}
